Remove `CKD_FORCE_INT128`
The conditional to use this macro caused int128 to not be used anyway,
however after fixing that the C++ implementation returns the wrong
results when T, U or V are int128. The C++ implementation for this type
likely received less care due to the static_assert rejecting int128 as a
non-integral type.